+++ l33tc0d3.cpp/709.cpp
@@ -39,7 +39,7 @@ namespace l33tc0d3cpp
 		//}
 
 		string toLowerCase(string str) {
-			for (int i = 0; i < str.length(); i++)
+			for (int i = 0; i < static_cast<int>(str.length()); i++)
 			{
 				auto&& ch = str[i];
